Recreated branch - (Task #19)
---
Task #19: Put Table and Chart into one row in the ReliabilityAnalysis
and AvailablityAnalaysis
