+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/model/RecoveryAutomatonTest.java
@@ -28,8 +28,6 @@ import de.dlr.sc.virsat.project.resources.VirSatResourceSet;
 // * Import Statements
 // *****************************************************************
 
-
-
 // *****************************************************************
 // * Class Declaration
 // *****************************************************************
