+++ de.dlr.sc.virsat.model.extension.fdir.ui/src-gen/de/dlr/sc/virsat/model/extension/fdir/ui/command/ACreateAddArrayElementAvailabilityCurveCommand.java
@@ -1,0 +1,36 @@
+/*******************************************************************************
+ * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package de.dlr.sc.virsat.model.extension.fdir.ui.command;
+
+import org.eclipse.emf.edit.domain.EditingDomain;
+import de.dlr.sc.virsat.model.dvlm.categories.ATypeInstance;
+import org.eclipse.emf.common.command.Command;
+import org.eclipse.emf.edit.command.AddCommand;
+import de.dlr.sc.virsat.model.dvlm.categories.propertyinstances.PropertyinstancesPackage;
+import de.dlr.sc.virsat.model.dvlm.categories.util.CategoryInstantiator;
+import de.dlr.sc.virsat.model.dvlm.categories.Category;
+import de.dlr.sc.virsat.model.dvlm.categories.propertyinstances.ArrayInstance;
+
+
+/**
+ * Auto Generated Abstract Generator Gap Class
+ * 
+ * Don't Manually modify this class
+ * 
+ * 
+ * 
+ */	
+public abstract class ACreateAddArrayElementAvailabilityCurveCommand {
+	
+	public Command create(EditingDomain editingDomain, ArrayInstance arrayInstance, Category type) {
+		ATypeInstance ati = new CategoryInstantiator().generateInstance(arrayInstance, type);
+		return AddCommand.create(editingDomain, arrayInstance, PropertyinstancesPackage.Literals.ARRAY_INSTANCE__ARRAY_INSTANCES, ati);
+	}
+}
