+++ de.dlr.sc.virsat.model.extension.fdir.ui/src/de/dlr/sc/virsat/model/extension/fdir/ui/snippet/UiSnippetSectionAvailabilityAnalysis.java
@@ -9,47 +9,24 @@
  *******************************************************************************/
 package de.dlr.sc.virsat.model.extension.fdir.ui.snippet;
 
-import java.awt.Color;
-import java.awt.Font;
-
-import org.eclipse.core.databinding.DataBindingContext;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.emf.common.command.Command;
-import org.eclipse.emf.common.notify.AdapterFactory;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.edit.domain.EditingDomain;
 import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.events.SelectionListener;
-import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.forms.widgets.FormToolkit;
-import org.jfree.chart.ChartFactory;
-import org.jfree.chart.JFreeChart;
-import org.jfree.chart.plot.XYPlot;
-import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer;
-import org.jfree.chart.title.TextTitle;
-import org.jfree.data.xy.XYSeriesCollection;
-import org.jfree.experimental.chart.swt.ChartComposite;
 
-import de.dlr.sc.virsat.commons.ui.jface.viewer.ISeriesXYValueLabelProvider;
-import de.dlr.sc.virsat.commons.ui.jface.viewer.XYSplineChartViewer;
-import de.dlr.sc.virsat.model.concept.list.IBeanList;
-import de.dlr.sc.virsat.model.concept.types.property.BeanPropertyFloat;
 import de.dlr.sc.virsat.model.dvlm.categories.CategoryAssignment;
-import de.dlr.sc.virsat.model.dvlm.qudv.util.QudvUnitHelper;
 import de.dlr.sc.virsat.model.extension.fdir.model.AvailabilityAnalysis;
-import de.dlr.sc.virsat.model.extension.fdir.model.Fault;
-import de.dlr.sc.virsat.project.ui.contentProvider.VirSatFilteredWrappedTreeContentProvider;
-import de.dlr.sc.virsat.project.ui.contentProvider.VirSatTransactionalAdapterFactoryContentProvider;
-import de.dlr.sc.virsat.project.ui.labelProvider.VirSatTransactionalAdapterFactoryLabelProvider;
 import de.dlr.sc.virsat.uiengine.ui.editor.snippets.IUiSnippet;
 
 /**
@@ -61,16 +38,6 @@ import de.dlr.sc.virsat.uiengine.ui.editor.snippets.IUiSnippet;
  * 
  */
 public class UiSnippetSectionAvailabilityAnalysis extends AUiSnippetSectionAvailabilityAnalysis implements IUiSnippet {
-	private static final int DEFAULT_CHART_HEIGHT = 600;
-	private static final int FONT_SIZE = 12;
-
-	private JFreeChart chart;
-	private ChartComposite chartComposite;
-	private Composite subSectionBody;
-
-	private XYSeriesCollection dataset;
-
-	private XYSplineChartViewer xyPlotChartViewer;
 
 	@Override
 	public void createSwt(FormToolkit toolkit, EditingDomain editingDomain, Composite composite, EObject initModel) {
@@ -101,7 +68,6 @@ public class UiSnippetSectionAvailabilityAnalysis extends AUiSnippetSectionAvail
 							return Status.CANCEL_STATUS;
 						}
 						editingDomain.getCommandStack().execute(availabilityAnalysisCommand);
-						Display.getDefault().syncExec(() -> xyPlotChartViewer.refresh());
 						monitor.done();
 						return Status.OK_STATUS;
 					}
@@ -110,138 +76,5 @@ public class UiSnippetSectionAvailabilityAnalysis extends AUiSnippetSectionAvail
 				job.schedule();
 			}
 		});
-
-		subSectionBody = toolkit.createComposite(composite);
-		layout = new GridLayout(1, true);
-		subSectionBody.setLayout(layout);
-		GridData gridDataAvailabilityChart = createDefaultGridData();
-		gridDataAvailabilityChart.horizontalSpan = 1;
-		gridDataAvailabilityChart.minimumHeight = DEFAULT_CHART_HEIGHT;
-		gridDataAvailabilityChart.heightHint = DEFAULT_CHART_HEIGHT;
-		subSectionBody.setLayoutData(gridDataAvailabilityChart);
-
-		// make a chart
-		chart = createChart();
-		chart.setBorderVisible(false);
-
-		chartComposite = new ChartComposite(subSectionBody, SWT.NONE, chart, true);
-		chartComposite.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-
-		xyPlotChartViewer = new XYSplineChartViewer(dataset, chartComposite);
-		xyPlotChartViewer.setContentProvider(new VirSatFilteredWrappedTreeContentProvider(
-				new VirSatTransactionalAdapterFactoryContentProvider(adapterFactory).setUpdateCaContainerLabels(true)) {
-			@Override
-			public Object[] getElements(Object inputElement) {
-				AvailabilityAnalysis relAnalysis = new AvailabilityAnalysis((CategoryAssignment) inputElement);
-				return new Object[] { relAnalysis };
-			}
-		});
-
-		xyPlotChartViewer.setLabelProvider(new AvailabilityChartLabelProvider(adapterFactory));
-		xyPlotChartViewer.setInput(model);
-	}
-
-	private static final double TO_PERCENT = 100;
-
-	/**
-	 * Creates the line chart
-	 * 
-	 * @return a configured line chart using the passed data set
-	 */
-	private JFreeChart createChart() {
-		AvailabilityAnalysis availabilityAnalysis = new AvailabilityAnalysis((CategoryAssignment) model);
-		String missionTimeUnit = availabilityAnalysis.getRemainingMissionTimeBean().getTypeInstance().getUnit()
-				.getSymbol();
-
-		dataset = new XYSeriesCollection();
-		JFreeChart chart = ChartFactory.createXYLineChart(null, "Mission Time [" + missionTimeUnit + "]",
-				"Availability [%]", dataset);
-
-		XYPlot plot = chart.getXYPlot();
-		plot.setNoDataMessage("No data available");
-		plot.setBackgroundPaint(Color.WHITE);
-		plot.setOutlineVisible(false);
-		plot.setBackgroundPaint(Color.white);
-		plot.setDomainGridlinePaint(Color.GRAY);
-		plot.setRangeGridlinePaint(Color.GRAY);
-		chart.setAntiAlias(true);
-		chart.setBorderVisible(true);
-		chart.setBackgroundPaint(Color.white);
-		plot.getRangeAxis().setRange(0, TO_PERCENT);
-
-		XYLineAndShapeRenderer renderer = new XYLineAndShapeRenderer(true, true);
-		plot.setRenderer(renderer);
-		renderer.setBaseShapesVisible(false);
-
-		// set the title
-		TextTitle myChartTitel = new TextTitle("Availability / Mission Success [%]",
-				new Font("SansSerif", Font.BOLD, FONT_SIZE));
-		chart.setTitle(myChartTitel);
-
-		return chart;
-	}
-
-	@Override
-	public void setDataBinding(DataBindingContext dbCtx, EditingDomain editingDomain, EObject model) {
-		super.setDataBinding(dbCtx, editingDomain, model);
-		xyPlotChartViewer.setInput(model);
-	}
-
-	/**
-	 * 
-	 * @author yoge_re
-	 *
-	 */
-	private class AvailabilityChartLabelProvider extends VirSatTransactionalAdapterFactoryLabelProvider
-			implements ISeriesXYValueLabelProvider {
-
-		/**
-		 * default constructor
-		 * 
-		 * @param adapterFactory
-		 *            the adapter Factory
-		 */
-		AvailabilityChartLabelProvider(AdapterFactory adapterFactory) {
-			super(adapterFactory);
-		}
-
-		@Override
-		public String getSeries(Object object) {
-			AvailabilityAnalysis availAnalysis = (AvailabilityAnalysis) object;
-			Fault fault = availAnalysis.getParentCaBeanOfClass(Fault.class);
-			if (fault != null) {
-				return "Availability of " + fault.getName();
-			} else {
-				return "Please add a fault";
-			}
-		}
-
-		@Override
-		public Double[] getValuesX(Object object) {
-			AvailabilityAnalysis availAnalysis = (AvailabilityAnalysis) object;
-			double timestep = availAnalysis.getTimestepBean().getValueToBaseUnit();
-			double maxTime = availAnalysis.getRemainingMissionTime();
-			double delta = QudvUnitHelper.getInstance().convertFromBaseUnitToTargetUnit(availAnalysis.getRemainingMissionTimeBean().getTypeInstance().getUnit(), timestep);
-			int steps = (int) (maxTime / delta);
-			Double[] timeSteps = new Double[steps + 1];
-			for (int time = 0; time <= steps; ++time) {
-				timeSteps[time] = time * delta;
-			}
-	
-			XYPlot plot = chart.getXYPlot();
-			plot.getDomainAxis().setUpperBound(maxTime);
-
-			return timeSteps;
-		}
-
-		@Override
-		public Double[] getValuesY(Object object) {
-			AvailabilityAnalysis availAnalysis = (AvailabilityAnalysis) object;
-			IBeanList<BeanPropertyFloat> availabilityCurve = availAnalysis.getPointAvailabilityCurve();
-			Double[] availabilityValues = availabilityCurve.stream().mapToDouble(beanFloat -> beanFloat.getValue())
-					.boxed().toArray(Double[]::new);
-			return availabilityValues;
-		}
-
 	}
 }
