+++ de.dlr.sc.virsat.model.extension.fdir/concept/concept_v1_1.xmi
@@ -172,7 +172,7 @@
     <properties xsi:type="dvlm_cppd:FloatProperty" name="remainingMissionTime" quantityKindName="Time" unitName="Hour"/>
     <properties xsi:type="dvlm_cppd:FloatProperty" name="timestep" quantityKindName="Time" unitName="Hour" defaultValue="0.1"/>
     <properties xsi:type="dvlm_cppd:FloatProperty" name="steadyStateAvailability" quantityKindName="Dimensionless" unitName="Percent"/>
-    <properties xsi:type="dvlm_cppd:FloatProperty" name="pointAvailabilityCurve" quantityKindName="Dimensionless" unitName="Percent">
+    <properties xsi:type="dvlm_cppd:FloatProperty" name="availabilityCurve" quantityKindName="Dimensionless" unitName="Percent">
       <arrayModifier xsi:type="dvlm_cppd:DynamicArrayModifier"/>
     </properties>
   </categories>
