+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/model/RecoveryAutomaton.java
@@ -32,7 +32,7 @@ import de.dlr.sc.virsat.model.extension.fdir.recovery.minimizer.ComposedMinimize
  * 
  * 
  * 
- */	
+ */
 public class RecoveryAutomaton extends ARecoveryAutomaton {
 
 	/**
@@ -43,38 +43,47 @@ public class RecoveryAutomaton extends ARecoveryAutomaton {
 	}
 
 	/**
-	 * Constructor of Concept Class which will instantiate 
-	 * a CategoryAssignment in the background from the given concept
-	 * @param concept the concept where it will find the correct Category to instantiate from
+	 * Constructor of Concept Class which will instantiate a CategoryAssignment in
+	 * the background from the given concept
+	 * 
+	 * @param concept
+	 *            the concept where it will find the correct Category to instantiate
+	 *            from
 	 */
 	public RecoveryAutomaton(Concept concept) {
 		super(concept);
-	}	
+	}
 
 	/**
-	 * Constructor of Concept Class that can be initialized manually by a given Category Assignment
-	 * @param categoryAssignment The category Assignment to be used for background initialization of the Category bean
+	 * Constructor of Concept Class that can be initialized manually by a given
+	 * Category Assignment
+	 * 
+	 * @param categoryAssignment
+	 *            The category Assignment to be used for background initialization
+	 *            of the Category bean
 	 */
 	public RecoveryAutomaton(CategoryAssignment categoryAssignment) {
 		super(categoryAssignment);
 	}
-	
+
 	/**
 	 * Converts this automaton into a string adhering to the .dot format.
+	 * 
 	 * @return a string representation of the automaton in the .dot format.
 	 */
 	public String toDot() {
 		StringBuilder sb = new StringBuilder();
 		sb.append("digraph ra {\n");
-		
+
 		for (State state : getStates()) {
 			sb.append(state.toString() + " [label=" + state.toString() + "]\n");
 		}
-		
+
 		for (Transition transition : getTransitions()) {
-			sb.append(transition.getFrom().toString()  + " -> " + transition.getTo().toString() + " [label=\"" + transition.toString() + "\"]\n");
+			sb.append(transition.getFrom().toString() + " -> " + transition.getTo().toString() + " [label=\""
+					+ transition.toString() + "\"]\n");
 		}
-		
+
 		sb.append("}");
 		return sb.toString();
 	}
